Added a "slug" b ased Page
Now the Pages are created using "slugs"

Problem remains:

Comments are NOT properly embedded into News post_type.

Some other problems:

I think it would be better to add Comments as a Plugin, but the difficulty is even higher.

So for next week, I will try to find out what the problem is with adding Comments.

Also, some other points:

-Settings
-Landing Page
-TBA